Split ExtraBlockData implementation from API
Extra block data is now computed eagerly rather than on-demand
