+++ library/blocks/block-extensions/src/main/java/org/quiltmc/qsl/block/extensions/api/data/package-info.java
@@ -1,5 +1,5 @@
 /**
- * <h2>{@linkplain org.quiltmc.qsl.block.extensions.api.data.ExtraBlockData Extra Block Data}</h2>
+ * <h2>{@linkplain org.quiltmc.qsl.block.extensions.api.data.ExtraBlockDataImpl Extra Block Data}</h2>
  * An extensible collection of data that is tied to a specific {@link net.minecraft.block.Block Block} instance.
  */
 package org.quiltmc.qsl.block.extensions.api.data;
