+++ library/blocks/block-extensions/src/main/resources/qsl_block_extensions.mixin.json
@@ -6,6 +6,7 @@
     "AbstractBlockAccessor",
     "AbstractBlockSettingsAccessor",
     "AbstractBlockSettingsMixin",
+    "BlockMixin",
     "MaterialBuilderAccessor"
   ],
   "client": [
